==> 2. Analysis Code/6. Making final results/3. data_average.ipynb <==
## **Code : Make variable-specific values average**
- Writer : Donghyeon Kim
- Update : 2022.11.15.
- 대상 : 태양광 사용 가구 & 미사용 가구
- 분석사항
    - 유의 : 가구별 데이터가 아닌, 사용/미사용 카테고리에서 평균치 산출하는 것
    1) 보정된 데이터 1시간 단위(사용/미사용)
    2) 보정된 데이터 1개월 단위(사용/미사용)

<hr>

## **Prior Settings**

# 라이브러리 설정
import os
import pandas as pd
import numpy as np
import openpyxl

# 사용자 함수 호출
# get_project_root : 파일의 상위-상위 경로 호출
# get_name_use_final : 태양광 사용자 이름 호출(최종)
# get_name_not_final : 태양광 미사용자 이름 호출(최종)
from pack_utils import get_project_root, get_name_use_final, get_name_not_final

### **Data**
#### Hour(1시간 단위)

# All Data
root = get_project_root()
folder_root = os.path.join(root, 'data_final')
file_name = os.path.join(folder_root, 'final_data_hour.xlsx')
df_all_hour = pd.read_excel(file_name)

# Filtering
# (1) Using Solar
df_all_hour_use = df_all_hour[df_all_hour.type == 'use']

# (2) Not Using Solar
df_all_hour_not = df_all_hour[df_all_hour.type == 'not']

df_all_hour.info()

df_all_hour_use.drop(columns=['id_hh', 'id_hs', 'owner']).head()

df_all_hour_not.drop(columns=['id_hh', 'id_hs', 'owner']).head()

<hr>

## **1시간 단위 : 평균치 구하기**
- 구분 : 태양광 사용 / 미사용
- 변수 : 태양광 발전 관련 변수 한정

### (1) Using Solar

# 결과 Dictionary
data_time = {}
data_time['연도'] = []
data_time['월'] = []
data_time['일'] = []
data_time['시간'] = []
data_time['발전량(kWh)'] = []
data_time['전력소비량(kWh)'] = []
data_time['수전전력량(kWh)'] = []
data_time['잉여전력량(kWh)'] = []
data_time['자가소비율'] = []
data_time['자가공급률'] = []
data_time['type'] = []

u_year = df_all_hour_use['연도'].unique().tolist()

for y in u_year:
    date_cond1 = (df_all_hour_use['연도'] == y)
    day_filter1 = df_all_hour_use[date_cond1]
    u_month = day_filter1['월'].unique().tolist()
    
    for m in u_month:
        date_cond2 = (day_filter1['월'] == m)
        day_filter2 = day_filter1[date_cond2]
        u_day = day_filter2['일'].unique().tolist()
        
        for d in u_day:
            date_cond3 = (day_filter2['일'] == d)
            day_filter3 = day_filter2[date_cond3]
            u_hour = sorted(day_filter3['시간'].unique().tolist())
            
            for h in u_hour:
                date_cond4 = (day_filter3['시간'] == h)
                day_filter4 = day_filter3[date_cond4]
                
                # Dictionary Filling
                # 연도
                data_time['연도'].append(y)
                
                # 월
                data_time['월'].append(m)
                
                # 일
                data_time['일'].append(d)
                
                # 시간
                data_time['시간'].append(h)
                
                # 발전량(kWh)
                power = np.mean(day_filter4['발전량(kWh)'])
                data_time['발전량(kWh)'].append(power)

                # 전력소비량(kWh)
                consum = np.mean(day_filter4['전력소비량(kWh)'])
                data_time['전력소비량(kWh)'].append(consum)

                # 수전전력량(kWh)
                grid_consum = np.mean(day_filter4['수전전력량(kWh)'])
                data_time['수전전력량(kWh)'].append(grid_consum)
                
                # 잉여전력량(kWh)
                export = np.mean(day_filter4['잉여전력량(kWh)'])
                data_time['잉여전력량(kWh)'].append(export)
                
                # 자가소비율(SCR)
                try:
                    scr = round((power - export)/power, 3)
                    data_time['자가소비율'].append(scr)
                except ZeroDivisionError:
                    data_time['자가소비율'].append(0)
                
                # 자가공급률(SSR)
                try:
                    ssr = round((power - export)/consum, 3)
                    data_time['자가공급률'].append(ssr)
                except ZeroDivisionError:
                    data_time['자가공급률'].append(0)
                
                # Type(사용)
                data_time['type'].append('use')
                    
    data_frame_time = pd.DataFrame(data_time)

data_frame_time.head(24)

### (2) Not using Solar

# 결과 Dictionary
data_time2 = {}
data_time2['연도'] = []
data_time2['월'] = []
data_time2['일'] = []
data_time2['시간'] = []
data_time2['발전량(kWh)'] = []
data_time2['전력소비량(kWh)'] = []
data_time2['수전전력량(kWh)'] = []
data_time2['잉여전력량(kWh)'] = []
data_time2['자가소비율'] = []
data_time2['자가공급률'] = []
data_time2['type'] = []

u_year = df_all_hour_not['연도'].unique().tolist()

for y in u_year:
    date_cond1 = (df_all_hour_not['연도'] == y)
    day_filter1 = df_all_hour_not[date_cond1]
    u_month = day_filter1['월'].unique().tolist()
    
    for m in u_month:
        date_cond2 = (day_filter1['월'] == m)
        day_filter2 = day_filter1[date_cond2]
        u_day = day_filter2['일'].unique().tolist()
        
        for d in u_day:
            date_cond3 = (day_filter2['일'] == d)
            day_filter3 = day_filter2[date_cond3]
            u_hour = sorted(day_filter3['시간'].unique().tolist())
            
            for h in u_hour:
                date_cond4 = (day_filter3['시간'] == h)
                day_filter4 = day_filter3[date_cond4]
                
                # Dictionary Filling
                # 연도
                data_time2['연도'].append(y)
                
                # 월
                data_time2['월'].append(m)
                
                # 일
                data_time2['일'].append(d)
                
                # 시간
                data_time2['시간'].append(h)
                
                # 발전량(kWh)
                data_time2['발전량(kWh)'].append(np.nan)

                # 전력소비량(kWh)
                consum = np.mean(day_filter4['전력소비량(kWh)'])
                data_time2['전력소비량(kWh)'].append(consum)

                # 수전전력량(kWh)
                grid_consum = np.mean(day_filter4['수전전력량(kWh)'])
                data_time2['수전전력량(kWh)'].append(grid_consum)
                
                # 잉여전력량(kWh)
                data_time2['잉여전력량(kWh)'].append(np.nan)
                
                # 자가소비율(SCR)
                data_time2['자가소비율'].append(np.nan)
                
                # 자가공급률(SSR)
                data_time2['자가공급률'].append(np.nan)
                
                # Type(사용)
                data_time2['type'].append('not')
                    
    data_frame_time2 = pd.DataFrame(data_time2)

data_frame_time2.head(24)

### (3) Making xlsx file
변수는 동일하므로 Merging을 실행할 수 있으나, 별개 데이터 생성이 주목적이므로 생략함.

xlsx_name = os.path.join(folder_root, 'final_data_hour_use_average.xlsx')
data_frame_time.to_excel(xlsx_name, sheet_name='hour_average', index=False)

xlsx_name2 = os.path.join(folder_root, 'final_data_hour_not_average.xlsx')
data_frame_time2.to_excel(xlsx_name2, sheet_name='hour_average', index=False)

<hr>

## **1개월 단위 : 평균치 구하기**
- 구분 : 태양광 사용 / 미사용
- 변수 : 태양광 발전 관련 변수 한정

### (1) Using Solar

# 결과 Dictionary
data_time = {}
data_time['연도'] = []
data_time['월'] = []
data_time['발전량(kWh)'] = []
data_time['전력소비량(kWh)'] = []
data_time['수전전력량(kWh)'] = []
data_time['잉여전력량(kWh)'] = []
data_time['자가소비율'] = []
data_time['자가공급률'] = []
data_time['type'] = []

u_year = df_all_hour_use['연도'].unique().tolist()

for y in u_year:
    date_cond1 = (df_all_hour_use['연도'] == y)
    day_filter1 = df_all_hour_use[date_cond1]
    u_month = day_filter1['월'].unique().tolist()
    
    for m in u_month:
        date_cond2 = (day_filter1['월'] == m)
        day_filter2 = day_filter1[date_cond2]
                
        # Dictionary Filling
        # 연도
        data_time['연도'].append(y)
        
        # 월
        data_time['월'].append(m)
        
        # 발전량(kWh)
        power = np.mean(day_filter2['발전량(kWh)'] * 24)
        data_time['발전량(kWh)'].append(power)
        
        # 전력소비량(kWh)
        consum = np.mean(day_filter2['전력소비량(kWh)'] * 24)
        data_time['전력소비량(kWh)'].append(consum)

        # 수전전력량(kWh)
        grid_consum = np.mean(day_filter2['수전전력량(kWh)'] * 24)
        data_time['수전전력량(kWh)'].append(grid_consum)
        
        # 잉여전력량(kWh)
        export = np.mean(day_filter2['잉여전력량(kWh)'] * 24)
        data_time['잉여전력량(kWh)'].append(export)
        
        # 자가소비율(SCR)
        try:
            scr = round((power - export)/power, 3)
            data_time['자가소비율'].append(scr)
        except ZeroDivisionError:
            data_time['자가소비율'].append(0)
        
        # 자가공급률(SSR)
        try:
            ssr = round((power - export)/consum, 3)
            data_time['자가공급률'].append(ssr)
        except ZeroDivisionError:
            data_time['자가공급률'].append(0)
                
        # Type(사용)
        data_time['type'].append('use')
                    
    data_frame_time = pd.DataFrame(data_time)

data_frame_time

### (2) Not using Solar

# 결과 Dictionary
data_time2 = {}
data_time2['연도'] = []
data_time2['월'] = []
data_time2['발전량(kWh)'] = []
data_time2['전력소비량(kWh)'] = []
data_time2['수전전력량(kWh)'] = []
data_time2['잉여전력량(kWh)'] = []
data_time2['자가소비율'] = []
data_time2['자가공급률'] = []
data_time2['type'] = []

u_year = df_all_hour_not['연도'].unique().tolist()

for y in u_year:
    date_cond1 = (df_all_hour_not['연도'] == y)
    day_filter1 = df_all_hour_not[date_cond1]
    u_month = day_filter1['월'].unique().tolist()
    
    for m in u_month:
        date_cond2 = (day_filter1['월'] == m)
        day_filter2 = day_filter1[date_cond2]
        
        # Dictionary Filling
        # 연도
        data_time2['연도'].append(y)
                
        # 월
        data_time2['월'].append(m)
                
        # 발전량(kWh)
        data_time2['발전량(kWh)'].append(np.nan)

        # 전력소비량(kWh)
        consum = np.mean(day_filter2['전력소비량(kWh)'] * 24)
        data_time2['전력소비량(kWh)'].append(consum)

        # 수전전력량(kWh)
        grid_consum = np.mean(day_filter2['수전전력량(kWh)'] * 24)
        data_time2['수전전력량(kWh)'].append(grid_consum)
                
        # 잉여전력량(kWh)
        data_time2['잉여전력량(kWh)'].append(np.nan)
                
        # 자가소비율(SCR)
        data_time2['자가소비율'].append(np.nan)
                
        # 자가공급률(SSR)
        data_time2['자가공급률'].append(np.nan)
                
        # Type(미사용)
        data_time2['type'].append('not')
        
    data_frame_time2 = pd.DataFrame(data_time2)

data_frame_time2

### (3) Making xlsx file
변수는 동일하므로 Merging을 실행할 수 있으나, 별개 데이터 생성이 주목적이므로 생략함.

xlsx_name = os.path.join(folder_root, 'final_data_month_use_average.xlsx')
data_frame_time.to_excel(xlsx_name, sheet_name='month_average', index=False)

xlsx_name2 = os.path.join(folder_root, 'final_data_month_not_average.xlsx')
data_frame_time2.to_excel(xlsx_name2, sheet_name='month_average', index=False)

<hr>

## **계절 단위 : 평균치 구하기**
- 구분 : 태양광 사용 / 미사용
- 변수 : 태양광 발전 관련 변수 한정

### (1) Using Solar

# 결과 Dictionary
data_time = {}
data_time['계절'] = []
data_time['발전량(kWh)'] = []
data_time['전력소비량(kWh)'] = []
data_time['수전전력량(kWh)'] = []
data_time['잉여전력량(kWh)'] = []
data_time['자가소비율'] = []
data_time['자가공급률'] = []
data_time['type'] = []

# 계절 필터링
use_spring = df_all_hour_use[(df_all_hour_use.ym == '2021/5') | (df_all_hour_use.ym == '2022/3') | (df_all_hour_use.ym == '2022/4')]
use_summer = df_all_hour_use[(df_all_hour_use.ym == '2021/6') | (df_all_hour_use.ym == '2021/7') | (df_all_hour_use.ym == '2021/8')]
use_autumn = df_all_hour_use[(df_all_hour_use.ym == '2021/9') | (df_all_hour_use.ym == '2021/10') | (df_all_hour_use.ym == '2021/11')]
use_winter = df_all_hour_use[(df_all_hour_use.ym == '2021/12') | (df_all_hour_use.ym == '2022/1') | (df_all_hour_use.ym == '2022/2')]

data_time['계절'] = ['spring', 'summer', 'autumn', 'winter']

# Dictionary Filling
for variable in [use_spring, use_summer, use_autumn, use_winter]:
    # 발전량(kWh)
    power = np.mean(variable['발전량(kWh)'] * 24)
    data_time['발전량(kWh)'].append(power)
        
    # 전력소비량(kWh)
    consum = np.mean(variable['전력소비량(kWh)'] * 24)
    data_time['전력소비량(kWh)'].append(consum)

    # 수전전력량(kWh)
    grid_consum = np.mean(variable['수전전력량(kWh)'] * 24)
    data_time['수전전력량(kWh)'].append(grid_consum)
        
    # 잉여전력량(kWh)
    export = np.mean(variable['잉여전력량(kWh)'] * 24)
    data_time['잉여전력량(kWh)'].append(export)
            
    # 자가소비율(SCR)
    try:
        scr = round((power - export)/power, 3)
        data_time['자가소비율'].append(scr)
    except ZeroDivisionError:
        data_time['자가소비율'].append(0)
        
    # 자가공급률(SSR)
    try:
        ssr = round((power - export)/consum, 3)
        data_time['자가공급률'].append(ssr)
    except ZeroDivisionError:
        data_time['자가공급률'].append(0)
                
    # Type(사용)
    data_time['type'].append('use')
    
data_frame_time = pd.DataFrame(data_time)

data_frame_time

### (2) Not using Solar

# 결과 Dictionary
data_time2 = {}
data_time2['계절'] = []
data_time2['발전량(kWh)'] = []
data_time2['전력소비량(kWh)'] = []
data_time2['수전전력량(kWh)'] = []
data_time2['잉여전력량(kWh)'] = []
data_time2['자가소비율'] = []
data_time2['자가공급률'] = []
data_time2['type'] = []

# 계절 필터링
not_spring = df_all_hour_not[(df_all_hour_not.ym == '2021/5') | (df_all_hour_not.ym == '2022/3') | (df_all_hour_not.ym == '2022/4')]
not_summer = df_all_hour_not[(df_all_hour_not.ym == '2021/6') | (df_all_hour_not.ym == '2021/7') | (df_all_hour_not.ym == '2021/8')]
not_autumn = df_all_hour_not[(df_all_hour_not.ym == '2021/9') | (df_all_hour_not.ym == '2021/10') | (df_all_hour_not.ym == '2021/11')]
not_winter = df_all_hour_not[(df_all_hour_not.ym == '2021/12') | (df_all_hour_not.ym == '2022/1') | (df_all_hour_not.ym == '2022/2')]

data_time2['계절'] = ['spring', 'summer', 'autumn', 'winter']

# Dictionary Filling
for variable in [not_spring, not_summer, not_autumn, not_winter]:      
    # 발전량(kWh)
    data_time2['발전량(kWh)'].append(np.nan)

    # 전력소비량(kWh)
    consum = np.mean(variable['전력소비량(kWh)'] * 24)
    data_time2['전력소비량(kWh)'].append(consum)

    # 수전전력량(kWh)
    grid_consum = np.mean(variable['수전전력량(kWh)'] * 24)
    data_time2['수전전력량(kWh)'].append(grid_consum)
                
    # 잉여전력량(kWh)
    data_time2['잉여전력량(kWh)'].append(np.nan)
                
    # 자가소비율(SCR)
    data_time2['자가소비율'].append(np.nan)
                
    # 자가공급률(SSR)
    data_time2['자가공급률'].append(np.nan)
                
    # Type(미사용)
    data_time2['type'].append('not')
    
data_frame_time2 = pd.DataFrame(data_time2)

data_frame_time2

### (3) Making xlsx file
변수는 동일하므로 Merging을 실행할 수 있으나, 별개 데이터 생성이 주목적이므로 생략함.

xlsx_name = os.path.join(folder_root, 'final_data_season_use_average.xlsx')
data_frame_time.to_excel(xlsx_name, sheet_name='season_average', index=False)

xlsx_name2 = os.path.join(folder_root, 'final_data_sesason_not_average.xlsx')
data_frame_time2.to_excel(xlsx_name2, sheet_name='sesason_average', index=False)

<hr>

## **1년 단위 : 평균치 구하기**
- 구분 : 태양광 사용 / 미사용
- 변수 : 태양광 발전 관련 변수 한정
- 1년 : 2021.05 ~ 2022.04

### (1) Using & Not using Solar

# 결과 Dictionary
data_time = {}
data_time['발전량(kWh)'] = []
data_time['전력소비량(kWh)'] = []
data_time['수전전력량(kWh)'] = []
data_time['잉여전력량(kWh)'] = []
data_time['자가소비율'] = []
data_time['자가공급률'] = []
data_time['type'] = []

# 1년 필터링(2021.05 ~ 2022.04)
use_year = df_all_hour_use[(df_all_hour_use.ym == '2021/5') | (df_all_hour_use.ym == '2021/6') | (df_all_hour_use.ym == '2021/7') |
                           (df_all_hour_use.ym == '2021/8') | (df_all_hour_use.ym == '2021/9') | (df_all_hour_use.ym == '2021/10') |
                           (df_all_hour_use.ym == '2021/11') | (df_all_hour_use.ym == '2021/12') | (df_all_hour_use.ym == '2022/1') |
                           (df_all_hour_use.ym == '2022/2') | (df_all_hour_use.ym == '2022/3') | (df_all_hour_use.ym == '2022/4')]

not_year = df_all_hour_not[(df_all_hour_not.ym == '2021/5') | (df_all_hour_not.ym == '2021/6') | (df_all_hour_not.ym == '2021/7') |
                           (df_all_hour_not.ym == '2021/8') | (df_all_hour_not.ym == '2021/9') | (df_all_hour_not.ym == '2021/10') |
                           (df_all_hour_not.ym == '2021/11') | (df_all_hour_not.ym == '2021/12') | (df_all_hour_not.ym == '2022/1') |
                           (df_all_hour_not.ym == '2022/2') | (df_all_hour_not.ym == '2022/3') | (df_all_hour_not.ym == '2022/4')]

# Dictionary Filling
for i, variable in enumerate([use_year, not_year]):
    # 발전량(kWh)
    power = np.mean(variable['발전량(kWh)'] * 24)
    data_time['발전량(kWh)'].append(power)
        
    # 전력소비량(kWh)
    consum = np.mean(variable['전력소비량(kWh)'] * 24)
    data_time['전력소비량(kWh)'].append(consum)

    # 수전전력량(kWh)
    grid_consum = np.mean(variable['수전전력량(kWh)'] * 24)
    data_time['수전전력량(kWh)'].append(grid_consum)
        
    # 잉여전력량(kWh)
    export = np.mean(variable['잉여전력량(kWh)'] * 24)
    data_time['잉여전력량(kWh)'].append(export)
            
    # 자가소비율(SCR)
    scr = round((power - export)/power, 3)
    data_time['자가소비율'].append(scr)
        
    # 자가공급률(SSR)
    ssr = round((power - export)/consum, 3)
    data_time['자가공급률'].append(ssr)
                
    # Type
    if i == 0:
        data_time['type'].append('use')
    else:
        data_time['type'].append('not')
    
data_result = pd.DataFrame(data_time)

data_result

### (2) Making xlsx file

xlsx_name = os.path.join(folder_root, 'final_data_year_average.xlsx')
data_result.to_excel(xlsx_name, sheet_name='year_average', index=False)
