==> 2. Analysis Code/6. Making final results/8. holiday_and_weekend.ipynb <==
## **Code : Add variable - holiday, weekend**
- Writer : Donghyeon Kim
- Update : 2022.11.18.
- 대상
    1. 3kW Only Dataset
    2. 3kW + 300W + 18kW Dataset(Total)

<hr>

## **Prior Settings**

# 라이브러리 설정
import os
import pandas as pd
import numpy as np
import openpyxl
from datetime import datetime, date
import holidays

# 사용자 함수 호출
# get_project_root : 파일의 상위-상위 경로 호출
from pack_utils import get_project_root

<hr>

## **Data : 3kW Only**

### 1. Rawdata

# All Data
root = get_project_root()
folder_root = os.path.join(root, 'data_final_keei')
file_name = os.path.join(folder_root, 'Dataset Hours 20221118 3kW Only.xlsx')
df_all_hour = pd.read_excel(file_name, sheet_name=0)

# Filtering
# (1) Using Solar
df_all_hour_use = df_all_hour[df_all_hour.type == 'use']

# (2) Not Using Solar
df_all_hour_not = df_all_hour[df_all_hour.type == 'not']

df_all_hour.info()

### 2. 시간평균

# Average Data
root = get_project_root()
folder_root = os.path.join(root, 'data_final_keei')
file_name = os.path.join(folder_root, 'Dataset Hours 20221118 3kW Only.xlsx')
df_avg_hour = pd.read_excel(file_name, sheet_name=1)

df_avg_hour.info()

<hr>

### **주말 및 공휴일 체크 함수**

#### 주말 여부 체크

def check_weekend(date):
    weekday = date.weekday()
    if weekday <= 4:
        return False
    if weekday > 4:
        return True

#### 공휴일 여부 체크

kr_holidays = holidays.KR()
def check_holiday(date):
    mydate = date
    if mydate in kr_holidays:
        return True
    if mydate not in kr_holidays:
        return False

<hr>

### **주말 및 공휴일 추가**

#### 1. Rawdata : 사용자 별로 추가

user_name = df_all_hour.owner.unique().tolist()
weekend_result = []
holiday_result = []

for i in range(len(user_name)):
    df_all_hour_filter = df_all_hour[df_all_hour.owner == user_name[i]]
    u_year = df_all_hour_filter['연도'].unique()
    
    for y in u_year:
        day_filter1 = df_all_hour_filter[df_all_hour_filter['연도'] == y]
        u_month = day_filter1['월'].unique()

        for m in u_month:
            day_filter2 = day_filter1[day_filter1['월'] == m]
            u_day = day_filter2['일'].unique()

            for d in u_day:
                day_filter3 = day_filter2[day_filter2['일'] == d]
                u_hour = day_filter3['시간'].unique()
                
                for h in u_hour:
                    day_filter4 = day_filter3[day_filter3['시간'] == h]
                    
                    if check_weekend(date(y, m, d)) == True:
                        weekend_result.append(1)
                    else:
                        weekend_result.append(0)
                    
                    if check_holiday(date(y, m, d)) == True:
                        holiday_result.append(1)
                    else:
                        holiday_result.append(0)
                        
df_all_hour['weekend'] = weekend_result
df_all_hour['holiday'] = holiday_result

#### 2. 시간평균 : 날짜 별로 추가

u_year = df_avg_hour['연도'].unique()
weekend_result = []
holiday_result = []
    
for y in u_year:
    day_filter1 = df_avg_hour[df_avg_hour['연도'] == y]
    u_month = day_filter1['월'].unique()

    for m in u_month:
        day_filter2 = day_filter1[day_filter1['월'] == m]
        u_day = day_filter2['일'].unique()

        for d in u_day:
            day_filter3 = day_filter2[day_filter2['일'] == d]
            u_hour = day_filter3['시간'].unique()
                
            for h in u_hour:
                day_filter4 = day_filter3[day_filter3['시간'] == h]
                    
                if check_weekend(date(y, m, d)) == True:
                    weekend_result.append(1)
                else:
                    weekend_result.append(0)
                    
                if check_holiday(date(y, m, d)) == True:
                    holiday_result.append(1)
                else:
                    holiday_result.append(0)

df_avg_hour['weekend'] = weekend_result
df_avg_hour['holiday'] = holiday_result

<hr>

### **엑셀 저장**

xlsx_name = os.path.join(folder_root, '3kW only result1.xlsx')
df_all_hour.to_excel(xlsx_name, sheet_name='RawData', index=False)

xlsx_name2 = os.path.join(folder_root, '3kW only result2.xlsx')
df_avg_hour.to_excel(xlsx_name2, sheet_name='시간평균', index=False)

<hr>

## **Data : 3kW + 300W + 18kW**

### 1. Rawdata

# All Data
root = get_project_root()
folder_root = os.path.join(root, 'data_final_keei')
file_name = os.path.join(folder_root, 'Datast Hours 20221118 3kW 300W 18kW.xlsx')
df_all_hour = pd.read_excel(file_name, sheet_name=0)

# Filtering
# (1) Using Solar
df_all_hour_use = df_all_hour[df_all_hour.type == 'use']

# (2) Not Using Solar
df_all_hour_not = df_all_hour[df_all_hour.type == 'not']

df_all_hour.info()

### 2. 시간평균

# Average Data
root = get_project_root()
folder_root = os.path.join(root, 'data_final_keei')
file_name = os.path.join(folder_root, 'Datast Hours 20221118 3kW 300W 18kW.xlsx')
df_avg_hour = pd.read_excel(file_name, sheet_name=1)

df_avg_hour.info()

<hr>

### **주말 및 공휴일 추가**

#### 1. Rawdata : 사용자 별로 추가

user_name = df_all_hour.owner.unique().tolist()
weekend_result = []
holiday_result = []

for i in range(len(user_name)):
    df_all_hour_filter = df_all_hour[df_all_hour.owner == user_name[i]]
    u_year = df_all_hour_filter['연도'].unique()
    
    for y in u_year:
        day_filter1 = df_all_hour_filter[df_all_hour_filter['연도'] == y]
        u_month = day_filter1['월'].unique()

        for m in u_month:
            day_filter2 = day_filter1[day_filter1['월'] == m]
            u_day = day_filter2['일'].unique()

            for d in u_day:
                day_filter3 = day_filter2[day_filter2['일'] == d]
                u_hour = day_filter3['시간'].unique()
                
                for h in u_hour:
                    day_filter4 = day_filter3[day_filter3['시간'] == h]
                    
                    if check_weekend(date(y, m, d)) == True:
                        weekend_result.append(1)
                    else:
                        weekend_result.append(0)
                    
                    if check_holiday(date(y, m, d)) == True:
                        holiday_result.append(1)
                    else:
                        holiday_result.append(0)
                        
df_all_hour['weekend'] = weekend_result
df_all_hour['holiday'] = holiday_result

#### 2. 시간평균 : 날짜 별로 추가

u_year = df_avg_hour['연도'].unique()
weekend_result = []
holiday_result = []
    
for y in u_year:
    day_filter1 = df_avg_hour[df_avg_hour['연도'] == y]
    u_month = day_filter1['월'].unique()

    for m in u_month:
        day_filter2 = day_filter1[day_filter1['월'] == m]
        u_day = day_filter2['일'].unique()

        for d in u_day:
            day_filter3 = day_filter2[day_filter2['일'] == d]
            u_hour = day_filter3['시간'].unique()
                
            for h in u_hour:
                day_filter4 = day_filter3[day_filter3['시간'] == h]
                    
                if check_weekend(date(y, m, d)) == True:
                    weekend_result.append(1)
                else:
                    weekend_result.append(0)
                    
                if check_holiday(date(y, m, d)) == True:
                    holiday_result.append(1)
                else:
                    holiday_result.append(0)

df_avg_hour['weekend'] = weekend_result
df_avg_hour['holiday'] = holiday_result

<hr>

### **엑셀 저장**

xlsx_name = os.path.join(folder_root, 'total result1.xlsx')
df_all_hour.to_excel(xlsx_name, sheet_name='RawData', index=False)

xlsx_name2 = os.path.join(folder_root, 'total result2.xlsx')
df_avg_hour.to_excel(xlsx_name2, sheet_name='시간평균', index=False)
